feat: Change sudoku module to sudoku row
Changed sudoku module (a box containing 9 sudoku boxes) to
a sudoku row (a row containing 9 sudoku boxes). This is a much
more effective / easier to work with change since we need to
run sudoku checking algorithms on the grid. In addition, changed
the comments to reflect the box to row change.